+++ pyriemann/transfer/_tools.py
@@ -2,30 +2,32 @@ import numpy as np
 
 
 def encode_domains(X, y, domain):
-    r"""Encode the domains of the matrices in the labels.
+    r"""Encode the domains of the data in the labels.
 
     We handle the possibility of having different domains for the datasets by
-    extending the labels of the matrices and including this information to
-    them. For instance, if we have a matrix X with class `left_hand` on the
-    `domain_01` then its extended label will be `domain_01/left_hand`. Note
-    that if the classes were integers at first, they will be converted to
+    extending the labels of the data and including this information to them.
+    For instance, if we have a datum X with class `left_hand` on the
+    `domain_01` then its extended label will be `domain_01/left_hand`.
+    Note that if the classes were integers at first, they will be converted to
     strings.
 
     Parameters
     ----------
-    X : ndarray, shape (n_matrices, n_channels, n_channels)
-        Set of SPD matrices.
-    y : ndarray, shape (n_matrices,)
-        Labels for each matrix.
-    domain : ndarray, shape (n_matrices,)
-        Domains for each matrix.
+    X : ndarray, shape (n_matrices, n_channels, n_channels) or \
+            shape (n_vectors, n_ts)
+        Set of SPD matrices or tangent vectors.
+    y : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Labels for each matrix or vector.
+    domain : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Domains for each matrix or vector.
 
     Returns
     -------
-    X_enc : ndarray, shape (n_matrices, n_channels, n_channels)
-        The same set of SPD matrices given as input.
-    y_enc : ndarray, shape (n_matrices,)
-        Extended labels for each matrix.
+    X_enc : ndarray, shape (n_matrices, n_channels, n_channels) or \
+            shape (n_vectors, n_ts)
+        The same data given as input.
+    y_enc : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Extended labels for each matrix or vector.
 
     See Also
     --------
@@ -38,33 +40,35 @@ def encode_domains(X, y, domain):
     if len(y) != len(domain):
         raise ValueError("Input lengths don't match")
 
-    y_enc = [str(d_) + '/' + str(y_) for (d_, y_) in zip(domain, y)]
+    y_enc = [str(d_) + "/" + str(y_) for (d_, y_) in zip(domain, y)]
     return X, np.array(y_enc)
 
 
 def decode_domains(X_enc, y_enc):
-    """Decode the domains of the matrices in the labels.
+    """Decode the domains of the data in the labels.
 
     We handle the possibility of having different domains for the datasets by
-    encoding the domain information into the labels of the matrices. This
-    method converts the data into its original form, with a separate data
+    encoding the domain information into the labels of the data.
+    This method converts the data into its original form, with a separate data
     structure for labels and for domains.
 
     Parameters
     ----------
-    X_enc : ndarray, shape (n_matrices, n_channels, n_channels)
-        Set of SPD matrices.
-    y_enc : ndarray, shape (n_matrices,)
-        Extended labels for each matrix.
+    X_enc : ndarray, shape (n_matrices, n_channels, n_channels) or \
+            shape (n_vectors, n_ts)
+        Set of SPD matrices or tangent vectors.
+    y_enc : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Extended labels for each matrix or vector.
 
     Returns
     -------
-    X : ndarray, shape (n_matrices, n_channels, n_channels)
-        Set of SPD matrices.
-    y : ndarray, shape (n_matrices,)
-        Labels for each matrix.
-    domain : ndarray, shape (n_matrices,)
-        Domains for each matrix.
+    X : ndarray, shape (n_matrices, n_channels, n_channels) or \
+            shape (n_vectors, n_ts)
+        The same data given as input.
+    y : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Labels for each matrix or vector.
+    domain : ndarray, shape (n_matrices,) or shape (n_vectors,)
+        Domains for each matrix or vector.
 
     See Also
     --------
@@ -76,7 +80,7 @@ def decode_domains(X_enc, y_enc):
     """
     y, domain = [], []
     for y_enc_ in y_enc:
-        y_dec_ = y_enc_.split('/')
+        y_dec_ = y_enc_.split("/")
         domain.append(y_dec_[-2])
         y.append(y_dec_[-1])
     return X_enc, np.array(y), np.array(domain)
@@ -86,9 +90,9 @@ class TLSplitter():
     """Class for handling the cross-validation splits of multi-domain data.
 
     This is a wrapper to sklearn's cross-validation iterators [1]_ which
-    ensures the handling of domain information with the data points. In fact,
+    ensures the handling of domain information with the data. In fact,
     the data from source domain is always fully available in the training
-    partition whereas the random splits are done on the data points from the
+    partition whereas the random splits are done on the data from the
     target domain.
 
     Parameters
@@ -96,18 +100,17 @@ class TLSplitter():
     target_domain : str
         Domain considered as target.
     cv : None | BaseCrossValidator | BaseShuffleSplit, default=None
-        An instance of a cross validation iterator from sklearn.
-
-    References
-    ----------
-    .. [1] https://scikit-learn.org/stable/modules/cross_validation.html#cross-validation-iterators
+        An instance of a cross-validation iterator from sklearn.
 
     Notes
     -----
     .. versionadded:: 0.4
+
+    References
+    ----------
+    .. [1] https://scikit-learn.org/stable/modules/cross_validation.html#cross-validation-iterators
     """  # noqa
     def __init__(self, target_domain, cv):
-
         self.target_domain = target_domain
         self.cv = cv
 
@@ -116,10 +119,11 @@ class TLSplitter():
 
         Parameters
         ----------
-        X : ndarray, shape (n_matrices, n_channels, n_channels)
-            Set of SPD matrices.
-        y : ndarray, shape (n_matrices,)
-            Extended labels for each matrix.
+        X : ndarray, shape (n_matrices, n_channels, n_channels) or \
+                shape (n_vectors, n_ts)
+            Set of SPD matrices or tangent vectors.
+        y : ndarray, shape (n_matrices,) or shape (n_vectors,)
+            Extended labels for each matrix or vector.
 
         Yields
         ------
@@ -129,24 +133,25 @@ class TLSplitter():
             The testing set indices for that split.
         """
 
-        # decode the domains of the data points
+        # decode the domains of the data
         X, y, domain = decode_domains(X, y)
 
-        # indentify the indices of the target dataset
+        # identify the indices of the target dataset
         idx_source = np.where(domain != self.target_domain)[0]
         idx_target = np.where(domain == self.target_domain)[0]
         y_target = y[idx_target]
 
-        # index of training-split for the target data points
+        # index of training-split for the target data
         ss_target = self.cv.split(idx_target, y_target)
         for train_sub_idx_target, test_sub_idx_target in ss_target:
             train_idx = np.concatenate(
-                [idx_source, idx_target[train_sub_idx_target]])
+                [idx_source, idx_target[train_sub_idx_target]]
+            )
             test_idx = idx_target[test_sub_idx_target]
             yield train_idx, test_idx
 
     def get_n_splits(self, X=None, y=None):
-        """Returns the number of splitting iterations in the cross-validator.
+        """Return the number of splitting iterations in the cross-validator.
 
         Parameters
         ----------
@@ -158,6 +163,6 @@ class TLSplitter():
         Returns
         -------
         n_splits : int
-            Returns the number of splitting iterations in the cross-validator.
+            Number of splitting iterations in the cross-validator.
         """
         return self.cv.n_splits